# Conflicts:
#	backend/BCAPIServer/src/main/java/BCServerMain.java

+++ backend/BCAPIServer/src/main/java/daos/User.java
@@ -84,7 +84,7 @@ public class User {
   private PreparedStatement likePostStatement;
 
   private static final String UNLIKE_POST =
-      "DELETE FROM liked_posts WHERE post_id = ?";
+      "DELETE FROM liked_posts WHERE user_id = ? AND post_id = ?";
   private PreparedStatement unlikePostStatement;
 
   private static final String GET_LIKED_POSTS =
@@ -97,6 +97,10 @@ public class User {
 
   private static final String GET_POSTS_BY_ID_STUB =
       "SELECT * FROM book_posts WHERE ";
+
+  private static final String DELETE_RECOMMENDATION_SQL =
+          "DELETE FROM sent_recommendations WHERE sender_username = ? AND recipient_username = ? AND book_key = ?";
+  private PreparedStatement deleteRecommendationStatement;
 
   /**
    * Creates a User class which allows user-specific interaction with the database. This class
@@ -139,6 +143,23 @@ public class User {
     rs.close();
 
     return userProfile;
+  }
+
+  public UserResult deleteRecommendation(String senderID, String bookKey) {
+    if (senderID.equals("") || bookKey.equals("")) {
+      return UserResult.INVALID;
+    }
+    try {
+      deleteRecommendationStatement.clearParameters();
+      deleteRecommendationStatement.setString(1, senderID);
+      deleteRecommendationStatement.setString(2, this.user);
+      deleteRecommendationStatement.setString(3, bookKey);
+      deleteRecommendationStatement.execute();
+    } catch (SQLException e) {
+      e.printStackTrace();
+      return UserResult.FAIL;
+    }
+    return UserResult.SUCCESS;
   }
 
   public UserResult updateUserProfile(String username, String bio, String thumbnail)
@@ -589,7 +610,7 @@ public class User {
     return count == 1;
   }
 
-  private int getNumLikesFromUser(String post_id) throws SQLException {
+  public int getNumLikesFromUser(String post_id) throws SQLException {
     countLikedPostsByPidStatement.clearParameters();
     countLikedPostsByPidStatement.setString(1, post_id);
     countLikedPostsByPidStatement.setString(2, user);
@@ -636,7 +657,8 @@ public class User {
     }
 
     unlikePostStatement.clearParameters();
-    unlikePostStatement.setString(1, post_id);
+    unlikePostStatement.setString(1, user);
+    unlikePostStatement.setString(2, post_id);
     int num_rows = unlikePostStatement.executeUpdate();
     return num_rows == 1 && posts.decreaseLikedPostByID(post_id) ? UserResult.SUCCESS : UserResult.FAIL;
   }
@@ -716,5 +738,7 @@ public class User {
     getLikedPostsStatement = conn.prepareStatement(GET_LIKED_POSTS);
 
     countLikedPostsByPidStatement = conn.prepareStatement(COUNT_LIKED_POSTS_BY_PID);
+
+    deleteRecommendationStatement = conn.prepareStatement(DELETE_RECOMMENDATION_SQL);
   }
 }
